# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "NomadCrew RESTful API with authentication and WebSocket support",
        "title": "NomadCrew Backend API",
        "contact": {
            "name": "NomadCrew Team",
            "url": "https://nomadcrew.uk",
            "email": "support@nomadcrew.uk"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "api.nomadcrew.uk",
    "basePath": "/v1",
    "paths": {
...
        "/trips/{id}/chat/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves messages for a trip's chat with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "List chat messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of messages to return (default 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of chat messages with pagination info",
                        "schema": {
                            "$ref": "#/definitions/types.ChatMessagePaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid trip ID or query parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User is not a member of this trip",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends a new message in a trip's chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Send chat message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message content",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ChatMessageCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created message details",
                        "schema": {
                            "$ref": "#/definitions/types.ChatMessage"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid trip ID or message content",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User is not a member of this trip",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/trips/{id}/chat/messages/{messageId}/reactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all reactions for a specific message in a trip's chat.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "List reactions for a message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "messageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reactions for the message",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.ChatMessageReactionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid trip ID or message ID",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User is not a member of this trip",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - Message not found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not implemented - This feature is not yet available",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds a reaction to a message in a trip's chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Add reaction to message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "messageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reaction details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ChatMessageReactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response with message",
                        "schema": {
                            "$ref": "#/definitions/docs.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid trip ID, message ID, or reaction",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User is not a member of this trip",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/trips/{id}/chat/messages/{messageId}/reactions/{reactionType}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a reaction from a message in a trip's chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Remove reaction from message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "messageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Reaction type",
                        "name": "reactionType",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response with message",
                        "schema": {
                            "$ref": "#/definitions/docs.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid trip ID, message ID, or reaction type",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User is not a member of this trip",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/trips/{tripId}/chat/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a paginated list of chat messages for a specific trip.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips-chat"
                ],
                "summary": "List messages for a trip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "tripId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Timestamp to fetch messages before (ISO 8601)",
                        "name": "before",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of messages to fetch (default 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of chat messages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.ChatMessage"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User not a member of this trip",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - Trip not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docs.ChatMessageReactionResponse": {
            "description": "Detailed information about a chat message reaction.",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2023-10-27T10:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "reaction-uuid-123"
                },
                "messageId": {
                    "type": "string",
                    "example": "message-uuid-456"
                },
                "reaction": {
                    "description": "The emoji character(s)",
                    "type": "string",
                    "example": "👍"
                },
                "userId": {
                    "type": "string",
                    "example": "user-uuid-789"
                }
            }
        },
...
        "docs.ErrorResponse": {
            "description": "Error information",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "error": {
                    "description": "Detailed error information",
                    "type": "string",
                    "example": "Field 'name' is required"
                },
                "message": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Invalid request parameters"
                }
            }
        },
...
        "github_com_NomadCrew_nomad-crew-backend_types.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "types.ChatMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "contentType": {
                    "description": "Added: e.g., \"text\", \"image_url\", \"system\"",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "groupId": {
                    "description": "Specific group within the trip, if applicable",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "isEdited": {
                    "type": "boolean"
                },
                "reactions": {
                    "description": "Added",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ChatMessageReaction"
                    }
                },
                "sender": {
                    "description": "Added for sender details",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.MessageSender"
                        }
                    ]
                },
                "tripId": {
                    "description": "Often the primary key for partitioning/lookup",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "types.ChatMessageCreateRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "tripId": {
                    "type": "string"
                }
            }
        },
        "types.ChatMessagePaginatedResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ChatMessageWithUser"
                    }
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}